--- controllers/animalControllers/index.js
@@ -1,7 +1,0 @@
-const addAnimal = require("./addAnimal");
-const removeAnimal = require("./removeAnimal");
-
-module.exports = {
-  addAnimal,
-  removeAnimal,
-};
