+++ middlewares/is-valid-id.js
@@ -10,4 +10,4 @@ const isValidId = (req, res, next) => {
   next();
 };
 
-module.exports = isValidId;
+module.exports = { isValidId };
