+++ routes/api/animals.js
@@ -1,11 +1,11 @@
 const express = require("express");
 
-const ctrl = require("../../controllers/animalControllers");
+const ctrl = require("../../controllers/animal-controllers");
 const { addAnimalJoiSchema } = require("../../db/animal-model");
 const { asyncWrapper } = require("../../helpers/api-helpers");
 const { authenticate } = require("../../middlewares/auth-middleware");
-const isValidId = require("../../middlewares/isValidId");
-const validateBody = require("../../middlewares/validateBody");
+const { isValidId } = require("../../middlewares/is-valid-id");
+const validateBody = require("../../middlewares/validate-body");
 
 const router = express.Router();
